Tests/LibWeb/Text/expected/abspos-relayout-insert-static-position.txt

live=0,25,40,15
reference=0,25,40,15
partial-layouts=0 full-layouts=1
FAIL
